% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean purpleair.R
\name{clean_PA}
\alias{clean_PA}
\title{Clean and Calibrate PurpleAir}
\usage{
clean_PA(PA_df, criteria = "manufacture", calibrate = TRUE)
}
\arguments{
\item{PA_df}{PurpleAir data frame, should be read in with fread (data.table) to preserve column names properly.}

\item{criteria}{Either "manufacture" (default) or "epa" for EPA cleaning criteria.}

\item{calibrate}{Whether or not you want to calibrate values using US EPA's correction curve. If TRUE (default), produces an additional column (pm2.5_cal) with calibrated values.}
}
\description{
Clean PurpleAir data using QA/QC methods from either the manufacturer or the EPA
}
\author{
Jonathan Liu
}
