% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download airnow.R
\name{download_airnow}
\alias{download_airnow}
\title{Download AirNow data}
\usage{
download_airnow(start_date, end_date, output_path = "./")
}
\arguments{
\item{start_date}{The starting date in "YYYY-MM-DD" format}

\item{end_date}{The ending date in "YYYY-MM-DD" format}

\item{output_path}{The output path}
}
\description{
function to download AirNow data, currently only works for dates in the same year
}
\examples{
download_airnow("2020-01-01", "2020-02-01")
download_airnow("2019-08-10", "2019-12-10")
}
