% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pems get station.R
\name{get_pems_id}
\alias{get_pems_id}
\title{Scrap PeMS Station IDs}
\usage{
get_pems_id(
  lat = 33.819318,
  lon = -117.918775,
  pages = 5,
  remdr,
  wait_time = 4,
  gps = F
)
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longitude}

\item{pages}{Number of pages to scrape from PeMS website. Default is 5 pages}

\item{remdr}{RSelenium remote driver from pems_login}

\item{wait_time}{Time to wait between page loads on the PeMS website. Default is 4 seconds. Adjust based on strength of internet connection.}

\item{gps}{boolean of whether or not you want GPS coordinates in addition to station IDs. Can be slow, default is FALSE (no GPS)}
}
\description{
Look up PeMS station IDs, given a longitude and latitude. Requires a PeMS account, a Docker instance, and running pems_login first.
}
\author{
Jonathan Liu
}
