% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read airnow.R
\name{read_airnow}
\alias{read_airnow}
\title{Read AirNow data}
\usage{
read_airnow(airnow, wide = F)
}
\arguments{
\item{airnow}{Character vector AirNow file name. Can be length 1 or longer.}

\item{wide}{Boolean of whether you want the data in wide (TRUE) or long (FALSE) format. Default is FALSE (long data)}
}
\description{
Simple function to read AQS data files downloaded by the function download_airnow, rename columns, and convert from wide to long data
}
